refactor: restructure Tic Tac Toe application by separating components and game logic
