+++ src/views/Board.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import Square from "./Square";
+
+export default function Board({ board, onSquareClick }) {
+  return (
+    <div className="board">
+      {board.map((value, index) => (
+        <Square
+          key={index}
+          value={value}
+          onClick={() => onSquareClick(index)}
+        />
+      ))}
+    </div>
+  );
+}
