+++ src/views/Square.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import "../App.css";
+
+export default function Square({ value, onClick }) {
+  return (
+    <button className="square" onClick={onClick}>
+      {value}
+    </button>
+  );
+}
